+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [1.13.8](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-1.13.7...@stream-io/video-react-bindings-1.13.8) (2026-02-20)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.43.0`
+
+## [1.13.7](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-1.13.6...@stream-io/video-react-bindings-1.13.7) (2026-02-16)
+
+### Dependency Updates
+
+- `@stream-io/video-client` updated to version `1.42.3`
+
 ## [1.13.6](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-1.13.5...@stream-io/video-react-bindings-1.13.6) (2026-02-13)
 
 ### Dependency Updates